==> 04_mouse_phos/01__filt_norm.Rmd <==
---
title: "01__filt_norm__2"
output: html_document
---


Background:

filter and normalise data




###Reading in data
```{r}
###Auxiliary
load("data/raw/m_p_trueraw_map_uniprot.RData")

###Read in:
phos_mouse_raw_data_all <- read.csv("data/raw/m_p_trueraw_py.csv", stringsAsFactors = FALSE)[, -1]

###Just replicates and standard name:
phos_mouse_raw_data_reps <- phos_mouse_raw_data_all[, c("standard_name",
colnames(phos_mouse_raw_data_all)[c(grep("CHOW", colnames(phos_mouse_raw_data_all)),
  grep("HFD", colnames(phos_mouse_raw_data_all)),
  grep("REV", colnames(phos_mouse_raw_data_all)))])]
##Standard name as rownames:
rownames(phos_mouse_raw_data_reps) <- phos_mouse_raw_data_reps[, 1]
phos_mouse_raw_data_reps <- phos_mouse_raw_data_reps[, -1]

phos_mouse_raw_data_all_quantified_mouse <- phos_mouse_raw_data_reps[which(rowSums(is.na(phos_mouse_raw_data_reps[, 1:72])) < 72), ]
```

###Rename and add columns
```{r}
###Rename
colnames(phos_mouse_raw_data_reps) <- gsub("CHOW",
                                           "CHOW_BAS",
                                           colnames(phos_mouse_raw_data_reps))
colnames(phos_mouse_raw_data_reps) <- gsub("HFD",
                                           "HFD_BAS",
                                           colnames(phos_mouse_raw_data_reps))
colnames(phos_mouse_raw_data_reps) <- gsub("REV",
                                           "REV_BAS",
                                           colnames(phos_mouse_raw_data_reps))
colnames(phos_mouse_raw_data_reps) <- gsub("BAS_INS",
                                           "INS",
                                           colnames(phos_mouse_raw_data_reps))


###Add phosmatched names
phos_mouse_raw_data_reps$phosmatched_standard_name <- ""
for (i in 1:nrow(phos_mouse_raw_data_reps)){
  
  phos_mouse_raw_data_reps$phosmatched_standard_name[i] <-
    m_p_trueraw_map_uniprot$correct_id[which(m_p_trueraw_map_uniprot$standard_name ==
                                               rownames(phos_mouse_raw_data_reps)[i])]
}
```


###Initial dealing with NAs:
```{r}
no_nas <- rowSums(is.na(phos_mouse_raw_data_reps))
hist(no_nas, breaks = 20)
###Cut more than 65 NAs:
phos_mouse_raw_data_filt <- phos_mouse_raw_data_reps[-which(no_nas > 65),]
```

###Median normalise:
```{r}
###Visualise before normalisation:
pdf("output/images/normalisation/boxplot_before_normalisation_mouse.pdf")
boxplot(phos_mouse_raw_data_filt[, 1:72], las = 2, pch = 20, cex = 0.5)
dev.off()

###Median-normalise:

#filt data
medians <- apply(phos_mouse_raw_data_filt[, 1:72], 2, FUN = median, na.rm = TRUE)
overall_median <- median(medians)
med_diffs <- medians - overall_median
phos_mouse_raw_data_filt_norm <- phos_mouse_raw_data_filt
phos_mouse_raw_data_filt_norm[, 1:72] <- sweep(phos_mouse_raw_data_filt[, 1:72], 2, med_diffs, FUN = "-")

#unfilt data
medians <- apply(phos_mouse_raw_data_reps[, 1:72], 2, FUN = median, na.rm = TRUE)
overall_median <- median(medians)
med_diffs <- medians - overall_median
phos_mouse_raw_data_norm <- phos_mouse_raw_data_reps
phos_mouse_raw_data_norm[, 1:72] <- sweep(phos_mouse_raw_data_reps[, 1:72], 2, med_diffs, FUN = "-")


###Visualise after normalisation (filt):
pdf("output/images/normalisation/boxplot_after_normalisation_mouse.pdf")
boxplot(phos_mouse_raw_data_filt_norm[, 1:72], las = 2, pch = 20, cex = 0.5)
dev.off()
```

###Give phos_mouse_raw_data_filt_norm phosmatched names and other columns
```{r}
phos_mouse_raw_data_norm$old_standard_name <- rownames(phos_mouse_raw_data_norm)
rownames(phos_mouse_raw_data_norm) <- phos_mouse_raw_data_norm$phosmatched_standard_name

phos_mouse_raw_data_norm$standard_name <- phos_mouse_raw_data_norm$phosmatched_standard_name
phos_mouse_raw_data_norm$gene <- sapply(phos_mouse_raw_data_norm$standard_name,
                                                               function(x) strsplit(x,
                                                                                    "_")[[1]][1])
phos_mouse_raw_data_norm$uniprot <- sapply(phos_mouse_raw_data_norm$standard_name,
                                                               function(x) strsplit(x,
                                                                                    "_")[[1]][2])
phos_mouse_raw_data_norm$site <- sapply(phos_mouse_raw_data_norm$standard_name,
                                                               function(x) strsplit(x,
                                                                                    "_")[[1]][3])
phos_mouse_raw_data_norm$multiplicity <- sapply(phos_mouse_raw_data_norm$standard_name,
                                                               function(x) strsplit(x,
                                                                                    "_")[[1]][4])
phos_mouse_raw_data_norm$uniprot_site <- 
  apply(phos_mouse_raw_data_norm[, c("uniprot",
                                                            "site")],
        1,
        function(x) paste(x, collapse = "_"))
```


###Export
```{r}
save(phos_mouse_raw_data_filt_norm,
     file = "output/data/phos_mouse_raw_data_filt_norm__1.RData")

save(phos_mouse_raw_data_norm,
     file = "output/data/phos_mouse_raw_data_norm__2.RData")
```
