==> 04_mouse_phos/03__imputation.Rmd <==
---
title: "03__imputation__1"
output: html_document
---


Background:

 imputation



###Initialise
```{r}
library(stringr)
library(dplyr)
library(reshape2)
```

Load in data
```{r}
load("data/intermediate/phos_mouse_raw_data_filt_norm__1.RData")
```



####Block imputation


####Assess how much imputation would be done

Set up
```{r}
####Criteria: 0 in one condition, >= 7 in  other condition


###Check with imputation code
#high_cap is the maximum number of NAs for the full condition
#low_cap is the minimum number of NAs for the missing condition
block_imputer_toimputeonly <- function(data, treatments, n_s, high_cap, low_cap){

###Create to_impute_m:
to_impute_m <- matrix(FALSE, ncol = length(treatments), nrow = nrow(data))
colnames(to_impute_m) <- treatments
rownames(to_impute_m) <- rownames(data)

na_sums_m <- matrix(0, ncol = length(treatments), nrow = nrow(data))
colnames(na_sums_m) <- treatments
rownames(na_sums_m) <- rownames(data)

data_is_na <- is.na(data)
for (k in 1:length(treatments)){
    na_sums_m[,k] <- rowSums(data_is_na[, (n_s*(k-1) + 1):(n_s*k)])
  }


##Loop over blocks:
for (k in 1:length(treatments)){

  l_inx <- n_s*(k-1) + 1
  r_inx <- n_s*k

  if (k %% 2 == 1){
    check_l_or_r <- 1
  } else {
    check_l_or_r <- -1
  }
  
  ##Loop over proteins:
  for (i in 1:nrow(data)){
    #Check <= low_cap (Note reversal of inequalities):
    if (na_sums_m[i, k] >= low_cap){
      
      #Check to left/right, >= high_cap
      if (na_sums_m[i, k + check_l_or_r] <= high_cap){
      
        counter_mean <- mean(as.numeric(data[i, (l_inx + n_s*check_l_or_r):(r_inx + n_s*check_l_or_r)]), na.rm = TRUE)
        if (na_sums_m[i, k] == n_s){
          block_mean <- counter_mean - 1
        } else {
              block_mean <- mean(as.numeric(data[i, l_inx:r_inx]), na.rm = TRUE)
        }
       
        #Check block_mean lower than counter_mean
        if (block_mean < counter_mean){
          
          #Assign to to_impute_m
          to_impute_m[i, k] <- TRUE
      }
    } 
  }
  }
  
}

return(to_impute_m)
}

phos_mouse_raw_data_filt_norm_toblockimpute <- 
  block_imputer_toimputeonly(data = phos_mouse_raw_data_filt_norm[, 1:72], 
                           treatments = c("CHOW_BAS", 
                                          "CHOW_INS",
                                          "HFD_BAS", 
                                          "HFD_INS",
                                          "REV_BAS", 
                                          "REV_INS"), 
                           n_s = 12, 
                           high_cap = 5, 
                           low_cap = 12)


###Numbers:

##In each condition:
colSums(phos_mouse_raw_data_filt_norm_toblockimpute)
```



###Imputation

Impute below lowest value
```{r}
###Impute by downshifting from the lowest quantified value
#This is Sean's idea
#Sample from N(lowest_value - down_shift, SD), where lowest_value is the lowest value for the whole ppeptide, and SD is the SD of that ppeptide in the quantified group. E.g. if I am to impute into CTRL_BAS, I take the SD of CTRL_INS for that pppeptide
#to_impute_m is boolean matrix where colnames are conditions (e.g. CTRL_BAS, CTRL_INS), and TRUE indicates that a conditon needs to be imputed into
#data is data (numeric only i.e. no extra columns)
block_imp_below_site_min <- function(to_impute_m,
                                     data,
                                     downshift = 1){
    
    ##Get n structure (number of ns in each condition)
    conditions <- colnames(to_impute_m)
    n_structure <- sapply(conditions,
                              function(x) length(grep(x,
                                                      colnames(data))))
    
    ##Calculate site SDs in each condition, and site mins across conditions
    site_mins <- apply(data,
                       1,
                       FUN = "min",
                       na.rm = TRUE)
    condition_site_SDs <- matrix(0,
                            ncol = length(conditions),
                            nrow = nrow(data))
    colnames(condition_site_SDs) <- conditions
    rownames(condition_site_SDs) <- rownames(data)
    
    for (j in 1:length(conditions)){
      
      condition_site_SDs[, j] <- apply(data[, grep(conditions[j],
                                              colnames(data))],
                                  1,
                                  FUN = "sd",
                                  na.rm = TRUE)
    }
    
    
    ##Make start index. This will be used once I loop over individual values
    start_index <- 0
    
    
    
    ##Impute by looping over conditions:
    for (j in 1:length(conditions)){
    
    
    ##Loop over rows:
    for (i in 1:nrow(data)){
      
      if (to_impute_m[i, j]){
        
        ##Loop over each value within that condition
        for (k in 1:n_structure[j]){
          
          ##Check if value missing
          if (is.na(data[i, start_index + k])){
            
            data[i, start_index + k] <- rnorm(1, 
                                              mean = site_mins[i] - downshift,
                                              sd = condition_site_SDs[i, j + (-1)^(1 + j %% 2)])
          }
        }
      }
    }
    
    #Update start index
    start_index <- start_index + n_structure[j]
    }
    
    ##Return
    return(data)
}



###Impute with downshift 1
phos_mouse_raw_data_filt_norm_imp_below_ds1 <- phos_mouse_raw_data_filt_norm
phos_mouse_raw_data_filt_norm_imp_below_ds1[, 1:72] <- block_imp_below_site_min(to_impute_m = phos_mouse_raw_data_filt_norm_toblockimpute,
                                                                     data = phos_mouse_raw_data_filt_norm[, 1:72],
                                                                     downshift = 1)
phos_mouse_raw_data_filt_norm_imp_below_ds1[, sapply(colnames(phos_mouse_raw_data_filt_norm_toblockimpute),
                                          function(x) paste(c(x, 
                                                              "_blockimp"),
                                                            collapse = ""))] <- phos_mouse_raw_data_filt_norm_toblockimpute
```



####Neighbour

###Run neighbour imputation
```{r}
###Rules:
##If >= 5, impute
treatments_vector <- c("CHOW_BAS",
                       "CHOW_INS",
                       "HFD_BAS",
                       "HFD_INS",
                       "REV_BAS",
                       "REV_INS")

###neighbour_imputer function
neighbour_imputer <- function(data){
  ##Make to_impute_m:
  to_impute_m <- matrix(FALSE, ncol = 6, nrow = nrow(data))
  colnames(to_impute_m) <- treatments_vector
  rownames(to_impute_m) <- rownames(data)
  na_sums_m <- matrix(0, ncol = 6, nrow = nrow(data))
  colnames(na_sums_m) <- treatments_vector
  rownames(na_sums_m) <- rownames(data)
  data_is_na <- is.na(data)
  for (k in 1:6){
    na_sums_m[,k] <- rowSums(data_is_na[, (12*(k-1) + 1):(12*k)])
  }
  to_impute_m <- na_sums_m <= 7
  
  ##Initialise output
  imputed_data <- as.matrix(data)
  
  ##Inputing:
  ##Loop over blocks:
  for (k in 1:6){
    l_inx <- 12*(k-1) + 1
    r_inx <- 12*k
    ##Loop over proteins:
    for (i in 1:nrow(imputed_data)){
      if (to_impute_m[i,k]){
        imp_mean <- mean(imputed_data[i, l_inx:r_inx], na.rm = TRUE)
        imp_sd <- sd(imputed_data[i, l_inx:r_inx], na.rm = TRUE)
        ##Loop over individuals:
        for (j in 0:11){
          if (is.na(imputed_data[i, l_inx + j])){
            imputed_data[i, l_inx + j] <- rnorm(1, imp_mean, imp_sd)
          }
        }
      }
    }
  }
  ##Return imputed_data and to_impute_m
  return_list <- list(imputed_data, to_impute_m)
  names(return_list) <- c("imputed_data", "imputed_marker")
  return(return_list)
}
```

Impute:
```{r}
n_imputed_output <- neighbour_imputer(phos_mouse_raw_data_filt_norm_imp_below_ds1[, 1:72])
n_imputed_tags <- n_imputed_output[[2]]
n_imputed_data <- n_imputed_output[[1]]
```


Copy into data
```{r}
###Make df
phos_mouse_raw_data_filt_norm_imp_below_ds1_neighbimp <- phos_mouse_raw_data_filt_norm_imp_below_ds1

###Copy over all data
for (i in 1:nrow(phos_mouse_raw_data_filt_norm_imp_below_ds1_neighbimp)){
  
  phos_mouse_raw_data_filt_norm_imp_below_ds1_neighbimp[i, 1:72] <- 
    n_imputed_data[which(rownames(n_imputed_data) ==
                                           rownames(phos_mouse_raw_data_filt_norm_imp_below_ds1_neighbimp)[i]), 
                          1:72]
}

###Revert for block imputed sites
block_imp_indices <- which(rowSums(phos_mouse_raw_data_filt_norm_imp_below_ds1_neighbimp[, grep("blockimp",
                                                                                                colnames(phos_mouse_raw_data_filt_norm_imp_below_ds1_neighbimp))]) > 0)


conditions <- c("CHOW_BAS", 
                "CHOW_INS",
                "HFD_BAS", 
                "HFD_INS",
                "REV_BAS", 
                "REV_INS")

for (i in 1:length(block_imp_indices)){
  
  for (j in 1:length(conditions)){
    
    if (phos_mouse_raw_data_filt_norm_imp_below_ds1_neighbimp[block_imp_indices[i], 
                                                              paste(c(conditions[j],
                                                                      "_blockimp"),
                                                                    collapse = "")] == TRUE){
      
      phos_mouse_raw_data_filt_norm_imp_below_ds1_neighbimp[block_imp_indices[i], 
                                                                  grep(conditions[j],
                                                                       colnames(phos_mouse_raw_data_filt_norm_imp_below_ds1_neighbimp[, 1:72]))] <- 
        phos_mouse_raw_data_filt_norm_imp_below_ds1[block_imp_indices[i],
                                                    grep(conditions[j],
                                                         colnames(phos_mouse_raw_data_filt_norm_imp_below_ds1[, 1:72]))]
                                                                    }
  }
}
```

###Export
```{r}
phos_mouse_data_filt_norm_imp <- phos_mouse_raw_data_filt_norm_imp_below_ds1_neighbimp
save(phos_mouse_data_filt_norm_imp,
     file = "output/data/phos_mouse_data_filt_norm_imp__1.RData")
```
