==> 04_mouse_phos/04__stats.Rmd <==
---
title: "04__stats__2"
output: html_document
---


Background:

t-tests

Add extra columns: diet to CHOW FC in BAS or INS, mean in each group



###Initialise
```{r}
library(limma)
library(dplyr)
library(reshape2)
```

Load in data
```{r}
load("data/intermediate/phos_mouse_data_filt_norm_imp__1.RData")
```


###Run stats

Stats:
```{r}
###Make two_sample_DE function
two_sample_DE <- function(data_group1_group2, group1_name, group2_name){
  ##Make group1_name_vs_group2_name
  group1_name_vs_group2_name <- paste(c(group1_name,
                                        "VS",
                                        group2_name),
                                      collapse = "_")
  
  ##Make design
  design <- cbind(group1_name = 1, 
                  group1_name_vs_group2_name = c(rep(0, length(grep(group1_name, colnames(data_group1_group2)))),
                                  rep(1, length(grep(group2_name, colnames(data_group1_group2))))))
  ##Fit
  fit <- lmFit(data_group1_group2, design)
  fit <- eBayes(fit)
  plotSA(fit)
  print(summary(decideTests(fit, method = "separate", adjust.method = "fdr",lfc = 0.58 )))
  return(fit)
}


###Run tests (JvG)
CHOW_fit <- two_sample_DE(phos_mouse_data_filt_norm_imp[, grep("CHOW", colnames(phos_mouse_data_filt_norm_imp[, 1:72]))], 
              "CHOW_BAS", 
              "CHOW_INS")
CHOW_toptable <- topTable(CHOW_fit, coef=2, adjust="fdr", number = Inf)
CHOW_sig <- CHOW_toptable[which(CHOW_toptable$adj.P.Val < 0.05 & abs(CHOW_toptable$logFC) > 0.58),]
dim(CHOW_sig) #319

HFD_fit <- two_sample_DE(phos_mouse_data_filt_norm_imp[, grep("HFD", colnames(phos_mouse_data_filt_norm_imp[, 1:72]))], 
              "HFD_BAS", 
              "HFD_INS")
HFD_toptable <- topTable(HFD_fit, coef=2, adjust="fdr", number = Inf)
HFD_sig <- HFD_toptable[which(HFD_toptable$adj.P.Val < 0.05 & abs(HFD_toptable$logFC) > 0.58),]
dim(HFD_sig) #203

REV_fit <- two_sample_DE(phos_mouse_data_filt_norm_imp[, grep("REV", colnames(phos_mouse_data_filt_norm_imp[, 1:72]))], 
              "REV_BAS", 
              "REV_INS")
REV_toptable <- topTable(REV_fit, coef=2, adjust="fdr", number = Inf)
REV_sig <- REV_toptable[which(REV_toptable$adj.P.Val < 0.05 & abs(REV_toptable$logFC) > 0.58),]
dim(REV_sig) #229


###Combine and add to phos_mouse
colnames(CHOW_toptable) <-  c("CHOW_logFC",
                             "CHOW_AvgExpr",
                             "CHOW_t",
                             "CHOW_p_val",
                             "CHOW_adj_p_val",
                             "CHOW_B")
colnames(HFD_toptable) <-  c("HFD_logFC",
                             "HFD_AvgExpr",
                             "HFD_t",
                             "HFD_p_val",
                             "HFD_adj_p_val",
                             "HFD_B")
colnames(REV_toptable) <-  c("REV_logFC",
                             "REV_AvgExpr",
                             "REV_t",
                             "REV_p_val",
                             "REV_adj_p_val",
                             "REV_B")
CHOW_HFD_toptable <- merge(CHOW_toptable,
                           HFD_toptable,
                           by = "row.names")
rownames(CHOW_HFD_toptable) <- CHOW_HFD_toptable[, 1]
CHOW_HFD_toptable <- CHOW_HFD_toptable[, -1]
CHOW_HFD_REV_toptable <- merge(CHOW_HFD_toptable,
                               REV_toptable,
                               by = "row.names")
rownames(CHOW_HFD_REV_toptable) <- CHOW_HFD_REV_toptable[, 1]
CHOW_HFD_REV_toptable <- CHOW_HFD_REV_toptable[, -1]

#Add to phos_mouse
CHOW_HFD_REV_toptable <- CHOW_HFD_REV_toptable[rownames(phos_mouse_data_filt_norm_imp), ]
phos_mouse_data_filt_norm_imp_stats <- cbind(phos_mouse_data_filt_norm_imp,
                                             CHOW_HFD_REV_toptable)

```




####Add extra columns
```{r}
####FC diet to CHOW

###Function: Makes model vs ctrl FCs
#Data is data, columns of format MODEL_BAS/TREATMENT_n e.g. CTRL_INS_4
#method is "median" or "mean"
#models are models e.g. CI_BAS, DEX_BAS
#control is name for control e.g. CTRL_BAS, CTRL_INS
model_vs_ctrl_logFC_maker <- function(data,
                                      method = "median",
                                      models,
                                      control){


    ##Set up data
    FC_data <- matrix(NA,
                      nrow = nrow(data),
                      ncol = length(models))
    rownames(FC_data) <- rownames(data)
    colnames(FC_data) <- sapply(models,
                                function(x) paste(c(x, "_FC"),
                                                  collapse = ""))
    
    ##Get CTRL medians or (means)
    CTRL_medians <- apply(data[, grep(paste(c(control,
                                              "_\\d+"),
                                            collapse = ""),
                                      colnames(data))],
                          1,
                          FUN = method,
                          na.rm = TRUE)
    
    ##Loop over models
    for (j in 1:length(models)){
      
      FC_data[, j] <- apply(data[, grep(paste(c(models[j],
                                              "_\\d+"),
                                            collapse = ""),
                                      colnames(data))],
                          1,
                          FUN = method,
                          na.rm = TRUE) -
        CTRL_medians
    }
    
    return(FC_data)
}


###Assign FCs
phos_mouse_data_filt_norm_imp_stats[, c("HFD_on_CHOW_BAS_FC",
           "REV_on_CHOW_BAS_FC")] <- 
  model_vs_ctrl_logFC_maker(data = phos_mouse_data_filt_norm_imp_stats,
                            models = c("HFD_BAS",
                                           "REV_BAS"),
                            control = "CHOW_BAS",
                            method = "mean")
phos_mouse_data_filt_norm_imp_stats[, c("HFD_on_CHOW_INS_FC",
           "REV_on_CHOW_INS_FC")] <- 
  model_vs_ctrl_logFC_maker(data = phos_mouse_data_filt_norm_imp_stats,
                            models = c("HFD_INS",
                                           "REV_INS"),
                            control = "CHOW_INS",
                            method = "mean")



####Mean in each condition
conditions <- c("CHOW_BAS",
                "CHOW_INS",
                "HFD_BAS",
                "HFD_INS",
                "REV_BAS",
                "REV_INS")
for (i in 1:length(conditions)){
  
  phos_mouse_data_filt_norm_imp_stats[, paste(c(conditions[i],
                                                "_mean"),
                                              collapse = "")] <-
    apply(phos_mouse_data_filt_norm_imp_stats[, grep(paste(c(conditions[i],
                                                             "_\\d+"),
                                                           collapse = ""),
                                                     colnames(phos_mouse_data_filt_norm_imp_stats))],
          1,
          FUN = "mean",
          na.rm = TRUE)
}
```


###Export
```{r}
save(phos_mouse_data_filt_norm_imp_stats,
     file = "output/data/phos_mouse_data_filt_norm_imp_stats__2.RData")
```
