==> 04_mouse_phos/06__reg_and_def.Rmd <==
---
title: "06__reg_and_def__2"
output: html_document
---


Background:

Assign regulated and defective sites




###Initialise
```{r}
library(ggplot2)
library(stringr)
library(dplyr)
library(reshape2)
source("..\\scripts/essential_visualisations_JvG.R")
source("..\\scripts/enrichment_analysis_JvG.R")
source("..\\scripts/3T3L1_mouse_IR__DF__202104_objects.R")
source("..\\scripts/3T3L1_mouse_IR__DF__202104_functions.R")
```

Load in data
```{r}
load("data/intermediate/phos_mouse_data_filt_norm_imp_stats_phosmatched__2.RData")
phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef <- phos_mouse_data_filt_norm_imp_stats_phosmatched
```


###Assign regulated and defective
```{r}
###Regulated
phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$CHOW_reg <- FALSE
phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$CHOW_reg[which(phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$CHOW_adj_p_val < 0.05 &
                                                                 abs(phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$CHOW_logFC) > 0.58)] <- TRUE
phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$HFD_reg <- FALSE
phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$HFD_reg[which(phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$HFD_adj_p_val < 0.05 &
                                                                 abs(phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$HFD_logFC) > 0.58)] <- TRUE
phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$REV_reg <- FALSE
phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$REV_reg[which(phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$REV_adj_p_val < 0.05 &
                                                                 abs(phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$REV_logFC) > 0.58)] <- TRUE


###Defective
#Criteria:
#CHOW adjpval < 0.05
#CHOW logFC > 0.58 / < -0.58
#HFD logFC < 0.58 / > -0.58
phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$HFD_defective <- FALSE
phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$HFD_defective[which(phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$CHOW_reg == TRUE &
                                                                      ((phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$CHOW_logFC > 0.58 & 
                                                                         phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$HFD_logFC < 0.58) |
                                                                      (phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$CHOW_logFC < -0.58 &
                                                                         phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$HFD_logFC > -0.58)))] <- TRUE

table(phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$HFD_defective)


###Defective with REV returned
#defective
#REV: adjpval < 0.05, |logFC| > 0.58 in same direction as CHOW
phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$HFD_defective_REV_returned <- FALSE
phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$HFD_defective_REV_returned[which(phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$HFD_defective == TRUE &
                                                                                          phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$CHOW_logFC > 0.58 &
                                                                                          phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$REV_adj_p_val < 0.05 &
                                                                                          phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$REV_logFC > 0.58)] <- TRUE
phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$HFD_defective_REV_returned[which(phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$HFD_defective == TRUE &
                                                                                          phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$CHOW_logFC < -0.58 &
                                                                                          phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$REV_adj_p_val < 0.05 &
                                                                                          phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef$REV_logFC < -0.58)] <- TRUE
```

###Make proc database and export
```{r}
phos_mouse_proc <- phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef
save(phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef,
     file = "output/data/phos_mouse_data_filt_norm_imp_stats_phosmatched_regdef__2.RData")
save(phos_mouse_proc,
     file = "output/data/phos_mouse_proc__2.RData")
```

####Preferential defect of dephosphorylation
```{r}
###Carve up regulated sites into up and down, and defective and not defective

fishers_DE_pathway(background_genes = 
                     rownames(phos_mouse_proc[which(phos_mouse_proc$CHOW_reg == TRUE), ]),
                   DE_genes = 
                     rownames(phos_mouse_proc[which(phos_mouse_proc$HFD_defective == TRUE), ]),
                   pathway_genes = 
                     rownames(phos_mouse_proc[which(phos_mouse_proc$CHOW_reg == TRUE &
                                                      phos_mouse_proc$CHOW_logFC < -0.58), ]),
                   alternative = "greater")
```

####Big numbers of regulated and defective
```{r}
phos_mouse_regdef_summary_rownames <- c("Regulated ppeptides",
                                        "Regulated psites",
                                        "Regulated pproteins",
                                        "Defective ppeptides",
                                        "Defective psites",
                                        "Defective pproteins",
                                        "Defective_rev_return ppeptides",
                                        "Defective_rev_return psites",
                                        "Defective_rev_return pproteins")
phos_mouse_regdef_summary <- matrix(0,
                                    nrow = length(phos_mouse_regdef_summary_rownames),
                                    ncol = 5)
rownames(phos_mouse_regdef_summary) <- phos_mouse_regdef_summary_rownames
colnames(phos_mouse_regdef_summary) <- c("Category",
                                         "Object",
                                         "Total",
                                         "Up",
                                         "Down")
phos_mouse_regdef_summary <- as.data.frame(phos_mouse_regdef_summary)
phos_mouse_regdef_summary$Category <- sapply(rownames(phos_mouse_regdef_summary),
                                             function(x) strsplit(x, " ")[[1]][1])
phos_mouse_regdef_summary$Object <- sapply(rownames(phos_mouse_regdef_summary),
                                             function(x) strsplit(x, " ")[[1]][2])







###Regulated
phos_mouse_regdef_summary[1, 3:5] <- c(length(which(phos_mouse_proc$CHOW_reg == TRUE)),
                                    length(which(phos_mouse_proc$CHOW_reg == TRUE & phos_mouse_proc$CHOW_logFC > 0.58)),
                                    length(which(phos_mouse_proc$CHOW_reg == TRUE & phos_mouse_proc$CHOW_logFC < -0.58)))
phos_mouse_regdef_summary[2, 3:5] <- c(length(unique(phos_mouse_proc$uniprot_site[which(phos_mouse_proc$CHOW_reg == TRUE)])),
                                    length(unique(phos_mouse_proc$uniprot_site[which(phos_mouse_proc$CHOW_reg == TRUE &
                                                   phos_mouse_proc$CHOW_logFC > 0.58)])),
                                    length(unique(phos_mouse_proc$uniprot_site[which(phos_mouse_proc$CHOW_reg == TRUE &
                                                   phos_mouse_proc$CHOW_logFC < -0.58)])))
phos_mouse_regdef_summary[3, 3:5] <- c(length(unique(phos_mouse_proc$uniprot[which(phos_mouse_proc$CHOW_reg == TRUE)])),
                                    length(unique(phos_mouse_proc$uniprot[which(phos_mouse_proc$CHOW_reg == TRUE &
                                              phos_mouse_proc$CHOW_logFC > 0.58)])),
                                    length(unique(phos_mouse_proc$uniprot[which(phos_mouse_proc$CHOW_reg == TRUE &
                                              phos_mouse_proc$CHOW_logFC < -0.58)])))


###Defective
phos_mouse_regdef_summary[4, 3:5] <- c(length(which(phos_mouse_proc$HFD_defective == TRUE)),
                                       length(which(phos_mouse_proc$HFD_defective == TRUE & phos_mouse_proc$CHOW_logFC > 0.58)),
                                       length(which(phos_mouse_proc$HFD_defective == TRUE & phos_mouse_proc$CHOW_logFC < -0.58)))
phos_mouse_regdef_summary[5, 3:5] <- c(length(unique(phos_mouse_proc$uniprot_site[which(phos_mouse_proc$HFD_defective == TRUE)])),
                                       length(unique(phos_mouse_proc$uniprot_site[which(phos_mouse_proc$HFD_defective == TRUE &
                                                   phos_mouse_proc$CHOW_logFC > 0.58)])),
                                       length(unique(phos_mouse_proc$uniprot_site[which(phos_mouse_proc$HFD_defective == TRUE &
                                                   phos_mouse_proc$CHOW_logFC < -0.58)])))
phos_mouse_regdef_summary[6, 3:5] <- c(length(unique(phos_mouse_proc$uniprot[which(phos_mouse_proc$HFD_defective == TRUE)])),
                                       length(unique(phos_mouse_proc$uniprot[which(phos_mouse_proc$HFD_defective == TRUE &
                                              phos_mouse_proc$CHOW_logFC > 0.58)])),
                                       length(unique(phos_mouse_proc$uniprot[which(phos_mouse_proc$HFD_defective == TRUE &
                                              phos_mouse_proc$CHOW_logFC < -0.58)])))

###Defective with REV returned
phos_mouse_regdef_summary[7, 3:5] <- c(length(which(phos_mouse_proc$HFD_defective_REV_returned == TRUE)),
                                       length(which(phos_mouse_proc$HFD_defective_REV_returned == TRUE & phos_mouse_proc$CHOW_logFC > 0.58)),
                                       length(which(phos_mouse_proc$HFD_defective_REV_returned == TRUE & phos_mouse_proc$CHOW_logFC < -0.58)))
phos_mouse_regdef_summary[8, 3:5] <- c(length(unique(phos_mouse_proc$uniprot_site[which(phos_mouse_proc$HFD_defective_REV_returned == TRUE)])),
                                       length(unique(phos_mouse_proc$uniprot_site[which(phos_mouse_proc$HFD_defective_REV_returned == TRUE &
                                                   phos_mouse_proc$CHOW_logFC > 0.58)])),
                                       length(unique(phos_mouse_proc$uniprot_site[which(phos_mouse_proc$HFD_defective_REV_returned == TRUE &
                                                   phos_mouse_proc$CHOW_logFC < -0.58)])))
phos_mouse_regdef_summary[9, 3:5] <- c(length(unique(phos_mouse_proc$uniprot[which(phos_mouse_proc$HFD_defective_REV_returned == TRUE)])),
                                       length(unique(phos_mouse_proc$uniprot[which(phos_mouse_proc$HFD_defective_REV_returned == TRUE &
                                              phos_mouse_proc$CHOW_logFC > 0.58)])),
                                       length(unique(phos_mouse_proc$uniprot[which(phos_mouse_proc$HFD_defective_REV_returned == TRUE &
                                              phos_mouse_proc$CHOW_logFC < -0.58)])))


###Export
write.csv(phos_mouse_regdef_summary,
          file = "output/data/regdef/phos_mouse_regdef_summary__1.csv")
```


####Visualise defect
```{r}
####Exact numbers barplot
phos_mouse_regdef_summary_exact <- phos_mouse_regdef_summary[c("Regulated ppeptides",
                                                               "Defective ppeptides"), 
                                                             c("Category",
                                                               "Up",
                                                               "Down")]
phos_mouse_regdef_summary_exact[1, 2:3] <- phos_mouse_regdef_summary_exact[1, 2:3] -
  phos_mouse_regdef_summary_exact[2, 2:3]
phos_mouse_regdef_summary_exact_melt <- melt(phos_mouse_regdef_summary_exact)
phos_mouse_regdef_summary_exact_melt$Category <- factor(phos_mouse_regdef_summary_exact_melt$Category,
                                                        levels = c("Regulated",
                                                                   "Defective"))
levels(phos_mouse_regdef_summary_exact_melt$Category) <- c("Not defective",
                                                           "Defective")
levels(phos_mouse_regdef_summary_exact_melt$variable) <- c("up", "down")

##Plot
ggplot(phos_mouse_regdef_summary_exact_melt,
       aes(x = Category,
           y = value,
           fill = variable)) + 
  geom_col(position = "stack",
           width = 0.8) +
  scale_fill_manual(name = "Insulin Regulation",
                      labels = c("Upregulated",
                                 "Downregulated"),
                      values = up_down_colours) +
  theme(panel.grid = element_blank(),
        panel.background = element_blank(),
        axis.line = element_line(color = "black"),
        plot.title = element_text(hjust = 0.5),
        axis.text.y = element_text(colour = "black"),
        axis.text.x = element_text(colour = "black",
                                   angle = 90,
                                   hjust = 1,
                                   vjust = 0.5)) +
  labs(x = "Peptide type",
       y = "Number of peptides") +
  ggtitle("Insulin-regulated Peptides In Vivo") 
  ggsave("output/images/defective/defect_sites_up_down_exact_mouse.pdf",
         width = 2.5,
         height = 3)
```
