==> 04_mouse_phos/07__core_ins_sig.Rmd <==
---
title: "07__core_ins_sig__1"
output: html_document
---


Background:

Analyse sites on core insulin signalling proteins



###Initialise
```{r}
library(ggplot2)
library(stringr)
library(dplyr)
library(reshape2)

source("..\\scripts/3T3L1_mouse_IR__DF__202104_functions.R")
```

###Load in data
```{r}
load("data/intermediate/phos_mouse_proc__2.RData")
```


###Canonical insulin sig network

Read in, process, all sites and reg sites
```{r}
###Read in and process data

##Read in
can_ins_sig_proteins_EN <- readRDS("data/intermediate/ins_and_sean.rds")

##Decaps
can_ins_sig_proteins_EN <- sapply(can_ins_sig_proteins_EN,
                                  function(name){
                                    
                                    lower_name <- tolower(name)
                                    name_list <- strsplit(lower_name,
                                                          "")[[1]]
                                    name_list[1] <- toupper(name_list[1])
                                    final_name <- paste(name_list,
                                                        collapse = "")
                                  })

##All sites in mouse
can_ins_sig_mouse_sites_EN <- rownames(phos_mouse_proc[which(phos_mouse_proc$gene %in% 
                                                                            can_ins_sig_proteins_EN), ])

##All sites regulated in mouse
can_ins_sig_mouse_reg_sites_EN <- can_ins_sig_mouse_sites_EN[which(can_ins_sig_mouse_sites_EN %in%
                                                                     rownames(phos_mouse_proc[phos_mouse_proc$CHOW_reg == TRUE, ]))]
```

###Defect sites
```{r}
####Summaries and plots

##Summaries
phos_mouse_proc_reg_sites_summary_perc <- reg_sites_summariser_mouse(phos_mouse_proc,
                                                                      perc = TRUE)
colnames(phos_mouse_proc_reg_sites_summary_perc)[-1] <- sapply(colnames(phos_mouse_proc_reg_sites_summary_perc)[-1],
                                                               function(x) paste(c("reg_", x),
                                                                                 collapse = "")) 
phos_mouse_proc_reg_sites_summary_can_ins_perc <- reg_sites_summariser_mouse(phos_mouse_proc[can_ins_sig_mouse_sites_EN, ],
                                                                      perc = TRUE)
colnames(phos_mouse_proc_reg_sites_summary_can_ins_perc) <- sapply(colnames(phos_mouse_proc_reg_sites_summary_can_ins_perc),
                                                               function(x) paste(c("can_", x),
                                                                                 collapse = "")) 


##Density plot comparing num of defect sites to all ins reg

#Set up df
phos_mouse_proc_reg_sites_summary_comb_perc <- cbind(phos_mouse_proc_reg_sites_summary_perc,
                                                 phos_mouse_proc_reg_sites_summary_can_ins_perc[, -1])


###Dodged bar plot comparing perc of defect sites sites to all ins reg
phos_mouse_proc_reg_sites_summary_comb_perc_melt <- melt(phos_mouse_proc_reg_sites_summary_comb_perc[, c("site_group",
    "can_Total",
    "reg_Total")],
    id.vars = "site_group")
phos_mouse_proc_reg_sites_summary_comb_perc_melt$site_group <- factor(phos_mouse_proc_reg_sites_summary_comb_perc_melt$site_group,
                                                                      levels = c("Not defective",
                                                                                 "Defective"))
ggplot(data = phos_mouse_proc_reg_sites_summary_comb_perc_melt,
       aes(x = site_group,
           y = value,
           fill = variable)) +
  geom_col(position = position_dodge()) +
  scale_fill_manual(values = c("#414141",
                               "#848484"),
                    labels = c("Regulated on canonical proteins",
                               "All regulated"),
                    name = "Peptide group") +
    ggtitle("Insulin-regulated peptides overall and on canonical proteins") +
  theme(panel.background = element_blank(),
        panel.grid = element_blank(),
        axis.line = element_line(color = "black"),
        axis.text.y = element_text(size = 12,
                                 colour = "black"),
        axis.text.x = element_text(size = 12,
                                   colour = "black",
                                   angle = 90,
                                   hjust = 1,
                                   vjust = 0.5),
        axis.title = element_text(size = 14,
                                  hjust = 0.5)) +
  labs(x = "Site group",
       y = "Percentage of peptide group",
       size = 14) 
  ggsave("output/images/core_ins_sig/perc_barplot_olddefect_core_vs_bg_mouse.pdf",
         width = 4,
         height = 4)

##Export
write.csv(phos_mouse_proc_reg_sites_summary_comb_perc,
          file = "output/data/regdef/phos_mouse_proc_reg_sites_summary_comb_perc.csv")
```

Fisher's exact for lack of defect
```{r}
###Set up
reg_sites <- rownames(phos_mouse_proc[which(phos_mouse_proc$CHOW_reg == TRUE), ])
def_sites <- rownames(phos_mouse_proc[which(phos_mouse_proc$HFD_defective == TRUE), ])

k <- length(intersect(can_ins_sig_mouse_reg_sites_EN,
                                                def_sites))
K <- length(can_ins_sig_mouse_reg_sites_EN)
n <- length(def_sites)
N <- length(reg_sites)

contingency_table <- matrix(c(k,
                              K - k,
                              n - k,
                              N - n - K + k),
                            ncol = 2,
                            nrow = 2)

###Test
fisher.test(x = contingency_table,
            alternative = "less")


###Likelihoods of defect on core and non-core proteins, given regulated
#Core: 0.4305556
#Non-core: 0.6963563
length(intersect(def_sites,
                 can_ins_sig_mouse_reg_sites_EN))/length(can_ins_sig_mouse_reg_sites_EN)
length(setdiff(def_sites,
                 can_ins_sig_mouse_reg_sites_EN))/length(setdiff(reg_sites,
                                                                 can_ins_sig_mouse_reg_sites_EN))
```
