==> 04_mouse_phos/08__emergent.Rmd <==
---
title: "08__emergent__1"
output: html_document
---


Background:

Analyse sites with emergent phosphorylation in HFD



###Initialise
```{r}
library(reshape2)
library(ggplot2)
source("..\\scripts/essential_visualisations_JvG.R")
source("..\\scripts/enrichment_analysis_JvG.R")
source("..\\scripts/3T3L1_mouse_IR__DF__202104_objects.R")
source("..\\scripts/3T3L1_mouse_IR__DF__202104_functions.R")
```

###Load in data
```{r}
load("data/intermediate/phos_mouse_proc__2.RData")
load("data/intermediate/phos_mouse_raw_data_norm__2.RData")
```



###Filtering
```{r}
###Filtering
#CHOW adj_pval >= 0.05
#|CHOW_logFC| <= 0.58
#HFD adj_pval < 0.05
#|HFD_logFC| > 0.58


###Emergent nums
#49 up, 56 down
phos_mouse_proc$HFD_emergent_up <- FALSE
phos_mouse_proc$HFD_emergent_up[which(abs(phos_mouse_proc$CHOW_logFC) <= 0.58 &
                                                     phos_mouse_proc$CHOW_adj_p_val >= 0.05 &
                                                     phos_mouse_proc$HFD_logFC > 0.58 &
                                                     phos_mouse_proc$HFD_adj_p_val < 0.05)] <- TRUE
phos_mouse_proc$HFD_emergent_down <- FALSE
phos_mouse_proc$HFD_emergent_down[which(abs(phos_mouse_proc$CHOW_logFC) <= 0.58 &
                                                     phos_mouse_proc$CHOW_adj_p_val >= 0.05 &
                                                     phos_mouse_proc$HFD_logFC < -0.58 &
                                                     phos_mouse_proc$HFD_adj_p_val < 0.05)] <- TRUE

length(which(phos_mouse_proc$HFD_emergent_up == TRUE))
length(which(phos_mouse_proc$HFD_emergent_down == TRUE))


###Back down in rev
#37 up, 47 down
phos_mouse_proc$HFD_emergent_up_rev_back <- FALSE
phos_mouse_proc$HFD_emergent_up_rev_back[which(phos_mouse_proc$HFD_emergent_up == TRUE &
               abs(phos_mouse_proc$REV_logFC) <= 0.58 &
               phos_mouse_proc$REV_adj_p_val >= 0.05)] <- TRUE
phos_mouse_proc$HFD_emergent_down_rev_back <- FALSE
phos_mouse_proc$HFD_emergent_down_rev_back[which(phos_mouse_proc$HFD_emergent_down == TRUE &
               abs(phos_mouse_proc$REV_logFC) <= 0.58 &
               phos_mouse_proc$REV_adj_p_val >= 0.05)] <- TRUE

length(which(phos_mouse_proc$HFD_emergent_up_rev_back == TRUE))
length(which(phos_mouse_proc$HFD_emergent_down_rev_back == TRUE))

#Export
save(phos_mouse_proc,
     file = "output/data/phos_mouse_proc__3.RData")
```


###Visualise
```{r}
library(tidyverse)

#Summary bplot
#Defects
phos_mouse_regdef_summary_exact <- reg_sites_summariser_mouse(phos_mouse_proc)
#Add emergent
phos_mouse_regdef_summary_exact <- rbind(phos_mouse_regdef_summary_exact, 0)
phos_mouse_regdef_summary_exact[3, -1] <- c(sum(phos_mouse_proc$HFD_emergent_up | phos_mouse_proc$HFD_emergent_down),
  sum(phos_mouse_proc$HFD_emergent_up),
  sum(phos_mouse_proc$HFD_emergent_down))
phos_mouse_regdef_summary_exact[3, 1] <- "Emergent"
#Melt
phos_mouse_regdef_summary_exact_melt <- pivot_longer(phos_mouse_regdef_summary_exact[, c("site_group",
                                                                                         "Up",
                                                                                         "Down")],
                                                     cols = c("Up", "Down"),
                                                     names_to = "variable", values_to = "value") %>%
  mutate(variable = factor(variable, levels = c("Up", "Down")),
         Category = factor(site_group, levels = c("Not defective", "Defective", "Emergent")))
levels(phos_mouse_regdef_summary_exact_melt$variable) <- c("up", "down")

##Plot
ggplot(phos_mouse_regdef_summary_exact_melt,
       aes(x = Category,
           y = value,
           fill = variable)) + 
  geom_col(position = "stack",
           width = 0.8) +
  scale_fill_manual(name = "Insulin Regulation",
                      labels = c("Upregulated",
                                 "Downregulated"),
                      values = up_down_colours) +
  comfy_theme(include_xaxis = FALSE) +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5)) +
  labs(x = "Peptide type",
       y = "Number of peptides") +
  ggtitle("Insulin-regulated Peptides In Vivo") 
  ggsave_pdfpng("output/images/defective/bplot_defemerg_sites_up_down_exact_mouse",
         width = 2.2,
         height = 2.5)
  
#Export
write.csv(phos_mouse_regdef_summary_exact,
          file = "output/data/regdef/phos_mouse_regdef_summary_exact.csv")
```


Heatmaps
```{r}
##All sites

#wo pvals
plot(fc_sig_heatmap(data = phos_mouse_proc[which(phos_mouse_proc$HFD_emergent_up == TRUE |
                                                                       phos_mouse_proc$HFD_emergent_down == TRUE), ],
               fc_cols_wo_pvals = c("CHOW_logFC",
                           "HFD_logFC",
                           "REV_logFC"),
               order_column = "HFD_logFC",
               x_axis_names = c("CHOW",
                                "HFD",
                                "REV"),
               is_decreasing = FALSE)) +
  ggtitle("Mouse emergent sites") 
  ggsave("output/images/emergent/mouse_emergent_sites_nodots.pdf",
         height = 10,
         width = 4)
```



####Enrichment:

###GO:
```{r}
####BP
GO_BP_mouse_emergent_up <- as.data.frame(GO_enricher(background_genes =
                                                                 unique(phos_mouse_raw_data_norm$gene),
                  DE_genes = unique(phos_mouse_proc$gene[which(phos_mouse_proc$HFD_emergent_up == TRUE)]),
                  gene_label = "SYMBOL",
                  organism_database = org.Mm.eg.db,
                  ontology = "BP",
                  pathway_DE_intersection_threshold = 3))

GO_BP_mouse_emergent_down <- as.data.frame(GO_enricher(background_genes =
                                                                 unique(phos_mouse_raw_data_norm$gene),
                  DE_genes = unique(phos_mouse_proc$gene[which(phos_mouse_proc$HFD_emergent_down == TRUE)]),
                  gene_label = "SYMBOL",
                  organism_database = org.Mm.eg.db,
                  ontology = "BP",
                  pathway_DE_intersection_threshold = 3))

GO_BP_mouse_emergent_up_down <- as.data.frame(GO_enricher(background_genes =
                                                                 unique(phos_mouse_raw_data_norm$gene),
                  DE_genes = unique(phos_mouse_proc$gene[which(phos_mouse_proc$HFD_emergent_up == TRUE |
                                                                                   phos_mouse_proc$HFD_emergent_down == TRUE)]),
                  gene_label = "SYMBOL",
                  organism_database = org.Mm.eg.db,
                  ontology = "BP",
                  pathway_DE_intersection_threshold = 3))


####CC
GO_CC_mouse_emergent_up <- as.data.frame(GO_enricher(background_genes =
                                                                 unique(phos_mouse_raw_data_norm$gene),
                  DE_genes = unique(phos_mouse_proc$gene[which(phos_mouse_proc$HFD_emergent_up == TRUE)]),
                  gene_label = "SYMBOL",
                  organism_database = org.Mm.eg.db,
                  ontology = "CC",
                  pathway_DE_intersection_threshold = 3))

GO_CC_mouse_emergent_down <- as.data.frame(GO_enricher(background_genes =
                                                                 unique(phos_mouse_raw_data_norm$gene),
                  DE_genes = unique(phos_mouse_proc$gene[which(phos_mouse_proc$HFD_emergent_down == TRUE)]),
                  gene_label = "SYMBOL",
                  organism_database = org.Mm.eg.db,
                  ontology = "CC",
                  pathway_DE_intersection_threshold = 3))

GO_CC_mouse_emergent_up_down <- as.data.frame(GO_enricher(background_genes =
                                                                 unique(phos_mouse_raw_data_norm$gene),
                  DE_genes = unique(phos_mouse_proc$gene[which(phos_mouse_proc$HFD_emergent_up == TRUE |
                                                                                   phos_mouse_proc$HFD_emergent_down == TRUE)]),
                  gene_label = "SYMBOL",
                  organism_database = org.Mm.eg.db,
                  ontology = "CC",
                  pathway_DE_intersection_threshold = 3))


###Visualise


###Produce side-by-side plot of GO_CC_up and GO_BP_down
ggsave(pathway_enrichment_barplot_fuse(plot_df_list = list(GO_CC_mouse_emergent_up,
                                                    GO_BP_mouse_emergent_down),
                                       num_pathways = "sig_only",
                                       colours = c("red",
                                                   "blue"),
                                       sig_line = TRUE,
                                       sig_cutoff = 0.05,
                                       stack = "horizontal"),
       file = "output/images/emergent/enrichment/GO/GOCCupBBdown_mouse_emergent_down.pdf",
       width = 12,
       height = 2)

#Export
write.csv(GO_CC_mouse_emergent_up, file = "output/data/emergent/GO_CC_mouse_emergent_up.csv")
write.csv(GO_BP_mouse_emergent_down, file = "output/data/emergent/GO_BP_mouse_emergent_down.csv")
```
