==> 04_mouse_phos/09__GSK3.Rmd <==
---
title: "10__GSK3__1"
output: html_document
---


Background:

Analyse GSK3 in mouse




###Initialise
```{r}
library(ksea)
library(reshape2)
library(ggplot2)
source("..\\scripts/essential_visualisations_JvG.R")
source("..\\scripts/enrichment_analysis_JvG.R")
source("..\\scripts/boxplots_JvG.R")
source("..\\scripts/3T3L1_mouse_IR__DF__202104_objects.R")
source("..\\scripts/3T3L1_mouse_IR__DF__202104_functions.R")
```

###Load in data
```{r}
load("data/intermediate/phos_mouse_proc__2.RData")
load("data/intermediate/phos_gsk3i_proc__2.RData")
```


###Big numbers
```{r}
###Quantified in mouse and putative GSK3 substrate
#Ppeptides: 324
#Psites: 205
#Pproteins: 143
length(which(phos_mouse_proc$uniprot_site %in%
               phos_gsk3i_proc$uniprot_site[which(phos_gsk3i_proc$GSK3i_down_motif == TRUE)]))
length(unique(phos_mouse_proc$uniprot_site[which(phos_mouse_proc$uniprot_site %in%
               phos_gsk3i_proc$uniprot_site[which(phos_gsk3i_proc$GSK3i_down_motif == TRUE)])]))
length(unique(phos_mouse_proc$uniprot[which(phos_mouse_proc$uniprot_site %in%
               phos_gsk3i_proc$uniprot_site[which(phos_gsk3i_proc$GSK3i_down_motif == TRUE)])]))
```


###Visualise
```{r}
#Trarg1 S72
boxplot_mousephos(phos_mouse_proc,
                  "Trarg1_Q8C838_S72_1",
                  siglabels = TRUE) 
ggsave_pdfpng("output/images/gsk3/Trarg1_Q8C838_S72_1_simpler",
       width = 1.5,
       height = 2)
```


###KSEA
```{r, fig.show = "hide}
###Make dataframe for Gsk3i ppome that looks like psp_kinase_substrate
#So that the code works, I add in 2 copies of the same data (it needs there to be two or more kinases)
phos_gsk3i_kinase_substrate <- as.data.frame(cbind("GSK3",
                                rep(unique(phos_gsk3i_proc$uniprot_site[which(phos_gsk3i_proc$GSK3i_down_motif == TRUE)]),
    2)))
colnames(phos_gsk3i_kinase_substrate) <- c("KINASE",
                                           "uniprot_site")
phos_gsk3i_kinase_substrate$KINASE[(0.5*nrow(phos_gsk3i_kinase_substrate) + 1):nrow(phos_gsk3i_kinase_substrate)] <- "GSK3_copy"


###Prepare data
#Make colnames so it works
phos_mouse_proc$CHOW_FC <- phos_mouse_proc$CHOW_logFC
phos_mouse_proc$HFD_FC <- phos_mouse_proc$HFD_logFC
phos_mouse_proc$REV_FC <- phos_mouse_proc$REV_logFC
phos_mouse_proc$HFD_FC_on_CHOW_FC <- phos_mouse_proc$HFD_logFC - phos_mouse_proc$CHOW_logFC
phos_mouse_proc$REV_FC_on_CHOW_FC <- phos_mouse_proc$REV_logFC - phos_mouse_proc$CHOW_logFC


####Run enrichment

##Only with negative FC in CHOW
#INS/BAS
set.seed(287)
phos_mouse_ksea_INSBAS_output_CHOWnegFC <- ksea_beltrao_enricher_mult_conditions(phos_mouse_proc,
                                                             phos_gsk3i_kinase_substrate,
                                                             c("CHOW",
                                                                  "HFD",
                                                                  "REV"),
                                                             sub_search_data = 
                                                               phos_mouse_proc[which(phos_mouse_proc$CHOW_logFC < 0), ])

#INS/BAS normalised to CHOW
set.seed(21)
phos_mouse_ksea_INSBAS_output_CHOWnegFCnormtoCHOW <- ksea_beltrao_enricher_mult_conditions(phos_mouse_proc,
                                                             phos_gsk3i_kinase_substrate,
                                                             c("HFD_FC_on_CHOW",
                                                                  "REV_FC_on_CHOW"),
                                                             sub_search_data = 
                                                               phos_mouse_proc[which(phos_mouse_proc$CHOW_logFC < 0), ])
save(phos_mouse_ksea_INSBAS_output_CHOWnegFCnormtoCHOW,
     file = "output/data/gsk3/ksea/phos_mouse_ksea_INSBAS_output_CHOWnegFCnormtoCHOW__1.RData")



##DIET/CHOW, in BAS or INS

#Only with negative FC in COW

#BAS
set.seed(173)
phos_mouse_ksea_DIETCHOWBAS_output_CHOWnegFC <- ksea_beltrao_enricher_mult_conditions(phos_mouse_proc,
                                                             phos_gsk3i_kinase_substrate,
                                                             c("HFD_on_CHOW_BAS",
                                                                  "REV_on_CHOW_BAS"),
                                                             sub_search_data = 
                                                               phos_mouse_proc[which(phos_mouse_proc$CHOW_logFC < 0), ])

#INS
set.seed(111)
phos_mouse_ksea_DIETCHOWINS_output_CHOWnegFC <- ksea_beltrao_enricher_mult_conditions(phos_mouse_proc,
                                                             phos_gsk3i_kinase_substrate,
                                                             c("HFD_on_CHOW_INS",
                                                                  "REV_on_CHOW_INS"),
                                                             sub_search_data = 
                                                               phos_mouse_proc[which(phos_mouse_proc$CHOW_logFC < 0), ])



###Visualise

##INS/BAS

#Only with negative FC in CHOW
phos_mouse_ksea_INSBAS_output_output_CHOWnegFC_ES_melt <- melt(phos_mouse_ksea_INSBAS_output_CHOWnegFC[1, grep("ES",
                                                                       colnames(phos_mouse_ksea_INSBAS_output_CHOWnegFC))])
phos_mouse_ksea_INSBAS_output_output_CHOWnegFC_ES_melt$condition <- factor(gsub("_ES",
                                                    "",
                                                    phos_mouse_ksea_INSBAS_output_output_CHOWnegFC_ES_melt$variable),
                                               levels = c("CHOW",
                                                          "HFD",
                                                          "REV"))

ggplot(data = phos_mouse_ksea_INSBAS_output_output_CHOWnegFC_ES_melt,
       aes(x = condition,
           y = value,
           fill = condition)) +
  geom_col() +
  scale_fill_manual(values = SH_mouse_main_colours,
                    name = "Condition") +
  geom_hline(yintercept = 0) +
  comfy_theme() +
  theme(axis.text.x = element_text(colour = "black",
                                   angle = 90,
                                   hjust = 1,
                                   vjust = 0.5),
        axis.line.x = element_blank(),
        legend.position = "none") +
  labs(x = "Condition",
       y = "Enrichment score") +
  ggtitle("KSEA using GSK3i putative substrates (negCHOWFC)") 
  ggsave("output/images/gsk3/KSEA/Gsk3i_negCHOWFC_put_substrates_KSEA_mouse.pdf",
         width = 2.5,
         height = 2.5)



##DIET/CHOW 

#neg 
##Set up
phos_mouse_ksea_DIETCHOW_output_CHOWnegFC <- cbind(phos_mouse_ksea_DIETCHOWBAS_output_CHOWnegFC,
                                          phos_mouse_ksea_DIETCHOWINS_output_CHOWnegFC)
phos_mouse_ksea_DIETCHOWBAS_output_CHOWnegFC_melt <- cbind(melt(phos_mouse_ksea_DIETCHOW_output_CHOWnegFC[1, grep("ES",
                                             colnames(phos_mouse_ksea_DIETCHOW_output_CHOWnegFC))]),
                                             melt(phos_mouse_ksea_DIETCHOW_output_CHOWnegFC[1, grep("S_pval",
                                             colnames(phos_mouse_ksea_DIETCHOW_output_CHOWnegFC))]))
colnames(phos_mouse_ksea_DIETCHOWBAS_output_CHOWnegFC_melt) <- c("ES_var",
                                                     "ES",
                                                     "pval_var",
                                                     "pval")
phos_mouse_ksea_DIETCHOWBAS_output_CHOWnegFC_melt$condition <- gsub("FC_on_control_",
                                                        "",
                                                        as.character(phos_mouse_ksea_DIETCHOWBAS_output_CHOWnegFC_melt$ES_var))
phos_mouse_ksea_DIETCHOWBAS_output_CHOWnegFC_melt$condition <- gsub("_ES",
                                                        "",
                                                        phos_mouse_ksea_DIETCHOWBAS_output_CHOWnegFC_melt$condition)
phos_mouse_ksea_DIETCHOWBAS_output_CHOWnegFC_melt$condition <- factor(phos_mouse_ksea_DIETCHOWBAS_output_CHOWnegFC_melt$condition,
                                                          levels = unique(phos_mouse_ksea_DIETCHOWBAS_output_CHOWnegFC_melt$condition))
phos_mouse_ksea_DIETCHOWBAS_output_CHOWnegFC_melt$model <- 
  factor(sapply(as.character(phos_mouse_ksea_DIETCHOWBAS_output_CHOWnegFC_melt$condition),
                function(x) strsplit(x,
                                     "_")[[1]][1]),
         levels = c("HFD",
                    "REV"))


#Plot
ggplot() +
  geom_col(data = phos_mouse_ksea_DIETCHOWBAS_output_CHOWnegFC_melt,
           aes(x = condition,
               y = ES,
               colour = model),
           fill = "white",
           size = 0.4,
           width = 0.8) +
  geom_col(data = 
             phos_mouse_ksea_DIETCHOWBAS_output_CHOWnegFC_melt[which(phos_mouse_ksea_DIETCHOWBAS_output_CHOWnegFC_melt$pval < 0.05), ],
           aes(x = condition,
               y = ES,
               fill = model),
           width = 0.8) +
  geom_hline(yintercept = 0) +
  scale_colour_manual(values = SH_mouse_main_colours[-1],
                    name = "Diet") + 
  scale_fill_manual(values = SH_mouse_main_colours[-1],
                    name = "Diet") +
 comfy_theme() +
  theme(axis.text.x = element_text(colour = "black",
                                   angle = 90,
                                   hjust = 1,
                                   vjust = 0.5),
        axis.line.x = element_blank(),
        legend.position = "none") +
  labs(x = "Condition",
       y = "Enrichment score") +
  ggtitle("Putative GSK3 substrates") 
  ggsave("output/images/gsk3/KSEA/Gsk3i_negCHOWFC_put_substrates_KSEA_DIETCHOW_mouse.pdf",
         width = 2,
         height = 3)
  
#Output
write.csv(phos_mouse_ksea_INSBAS_output_CHOWnegFC,
          file = "output/data/Gsk3/ksea/phos_mouse_ksea_INSBAS_output_CHOWnegFC.csv")
write.csv(phos_mouse_ksea_INSBAS_output_CHOWnegFCnormtoCHOW,
          file = "output/data/Gsk3/ksea/phos_mouse_ksea_INSBAS_output_CHOWnegFCnormtoCHOW.csv")
write.csv(phos_mouse_ksea_DIETCHOW_output_CHOWnegFC,
          file = "output/data/Gsk3/ksea/phos_mouse_ksea_DIETCHOW_output_CHOWnegFC.csv")
```

###S9/S21
```{r}
#Simpler boxplot
boxplot_mousephos(phos_mouse_proc,
                  "Gsk3b_Q9WV60_S9_1",
                  siglabels = TRUE)
ggsave_pdfpng("output/images/gsk3/Gsk3b_Q9WV60_S9_1_simpler.pdf",
              width = 1.5, height = 2)


###t.tests

##BAS
t.test(as.numeric(phos_mouse_proc["Gsk3b_Q9WV60_S9_1", 
                                  grep("CHOW_BAS_\\d+",
                                       colnames(phos_mouse_proc))]),
       as.numeric(phos_mouse_proc["Gsk3b_Q9WV60_S9_1", 
                                  grep("HFD_BAS_\\d+",
                                       colnames(phos_mouse_proc))]))
t.test(as.numeric(phos_mouse_proc["Gsk3b_Q9WV60_S9_1", 
                                  grep("CHOW_BAS_\\d+",
                                       colnames(phos_mouse_proc))]),
       as.numeric(phos_mouse_proc["Gsk3b_Q9WV60_S9_1", 
                                  grep("REV_BAS_\\d+",
                                       colnames(phos_mouse_proc))]))

##INS
t.test(as.numeric(phos_mouse_proc["Gsk3b_Q9WV60_S9_1", 
                                  grep("CHOW_INS_\\d+",
                                       colnames(phos_mouse_proc))]),
       as.numeric(phos_mouse_proc["Gsk3b_Q9WV60_S9_1", 
                                  grep("HFD_INS_\\d+",
                                       colnames(phos_mouse_proc))]))
t.test(as.numeric(phos_mouse_proc["Gsk3b_Q9WV60_S9_1", 
                                  grep("CHOW_INS_\\d+",
                                       colnames(phos_mouse_proc))]),
       as.numeric(phos_mouse_proc["Gsk3b_Q9WV60_S9_1", 
                                  grep("REV_INS_\\d+",
                                       colnames(phos_mouse_proc))]))
```

###pY
Here I look at the proposed regulatory pYs
```{r}
library(purrr)
#Mouse
#Gsk3a
gsk3a_mouse_pY_args <- list(
  list(protein = "Gsk3a_Q2NL51_Y279_1"),
  list(protein = "Gsk3a_Q2NL51_Y279_2")
)
map(gsk3a_mouse_pY_args,
    function(x){
      boxplot_mousephos(phos_mouse_proc,
                  x$protein)
      ggsave_pdfpng(paste("output/images/gsk3/",
                          x$protein,
                          sep = ""),
                    width = 1.5, height = 2)
    })
```
